Refactor almost working, need to fix how errors appear
